Reorganizing tests, grouping base parser tests in on module, base v1 tests in another module and base csv tests in another module
